Move geometry unit tests to tests/ directory
